rockchip: spl: fix reloc gd and FDT blob pointer
When the Rockchip function dram_init() is called in the spl.c
board_init_f() the variable gd->ram_base returns 0. This is problematic
for functions after the gd relocation. Fix by adding ram.base to
gd->ram_base in sdram.c.

After the function spl_relocate_stack_gd is called in crt0.S the pointer
gd might be relocated and BSS with the FDT blob is erased. The pointer
gd->fdt_blob is no longer valid when enabled SPL_OF_REAL and SPL_MMC,
SPL_USB_HOST or SPL_USB_GADGET. This patch makes that FDT properties can
be parsed in the board_init_r() function.

Fix by setting this pointers simulair to board_f.c and copy the FDT to
it's new location at the end of DRAM.

Signed-off-by: Johan Jonker <jbx6244@gmail.com>